+++ gpudb/protocol/aggregate_unpivot.h
@@ -124,9 +124,17 @@ namespace gpudb
          *                      view this result table is part of.  The default
          *                      value is ''.
          *                              <li>
-         *                      gpudb::aggregate_unpivot_materialize_on_gpu: If
-         *                      @a true then the output columns will be cached
-         *                      on the GPU.
+         *                      gpudb::aggregate_unpivot_materialize_on_gpu: No
+         *                      longer used.  See <a
+         *                      href="../../rm/concepts.html"
+         *                      target="_top">Resource Management Concepts</a>
+         *                      for information about how resources are
+         *                      managed, <a href="../../rm/concepts.html"
+         *                      target="_top">Tier Strategy Concepts</a> for
+         *                      how resources are targeted for VRAM, and <a
+         *                      href="../../rm/usage.html#tier-strategies"
+         *                      target="_top">Tier Strategy Usage</a> for how
+         *                      to specify a table's priority in VRAM.
          *                      <ul>
          *                              <li> gpudb::aggregate_unpivot_true
          *                              <li> gpudb::aggregate_unpivot_false
@@ -250,9 +258,17 @@ namespace gpudb
          *                      view this result table is part of.  The default
          *                      value is ''.
          *                              <li>
-         *                      gpudb::aggregate_unpivot_materialize_on_gpu: If
-         *                      @a true then the output columns will be cached
-         *                      on the GPU.
+         *                      gpudb::aggregate_unpivot_materialize_on_gpu: No
+         *                      longer used.  See <a
+         *                      href="../../rm/concepts.html"
+         *                      target="_top">Resource Management Concepts</a>
+         *                      for information about how resources are
+         *                      managed, <a href="../../rm/concepts.html"
+         *                      target="_top">Tier Strategy Concepts</a> for
+         *                      how resources are targeted for VRAM, and <a
+         *                      href="../../rm/usage.html#tier-strategies"
+         *                      target="_top">Tier Strategy Usage</a> for how
+         *                      to specify a table's priority in VRAM.
          *                      <ul>
          *                              <li> gpudb::aggregate_unpivot_true
          *                              <li> gpudb::aggregate_unpivot_false