+++ gpudb/protocol/create_union.h
@@ -86,9 +86,17 @@ namespace gpudb
          *                      table will be a top-level table.  The default
          *                      value is ''.
          *                              <li>
-         *                      gpudb::create_union_materialize_on_gpu: If @a
-         *                      true, then the columns of the output table will
-         *                      be cached on the GPU.
+         *                      gpudb::create_union_materialize_on_gpu: No
+         *                      longer used.  See <a
+         *                      href="../../rm/concepts.html"
+         *                      target="_top">Resource Management Concepts</a>
+         *                      for information about how resources are
+         *                      managed, <a href="../../rm/concepts.html"
+         *                      target="_top">Tier Strategy Concepts</a> for
+         *                      how resources are targeted for VRAM, and <a
+         *                      href="../../rm/usage.html#tier-strategies"
+         *                      target="_top">Tier Strategy Usage</a> for how
+         *                      to specify a table's priority in VRAM.
          *                      <ul>
          *                              <li> gpudb::create_union_true
          *                              <li> gpudb::create_union_false